minor python 3 syntax updates
tabs/spaces and (Q+1)/2 ==> int( (Q+1)/2 ).... for some reason this used to return an integer and now needs to be explicitly told to return an integer... either that or iterations used to take floats happily and now need integers... in any event for Q=5 this returns 3 and not 3.

+++ misc/python_prototypes/2d_nesting/projection_step_details.py
@@ -246,7 +246,7 @@ w_exact=np.zeros_like(w); b_exact=np.zeros_like(b); p_exact=np.zeros_like(b)
 
 
 # locations for plotting slices
-K = (nz-1)/4  ;  I = (nx-1)/4
+K = int((nz-1)/4)  ;  I = int((nx-1)/4)
 
 
 #--------------------------------------------------------------------------------
